converts all maxtimes to ints

+++ mozilla/config.py
@@ -155,7 +155,7 @@ PLATFORM_VARS = {
                 'reboot_command': ['scripts/external_tools/count_and_reboot.py',
                                    '-f', '../reboot_count.txt', '-n', '1', '-z'],
                 'script_timeout': 3 * 3600,
-                'script_maxtime': 5.5 * 3600,
+                'script_maxtime': int(5.5 * 3600),
             },
             'dep_signing_servers': 'dep-signing',
             'base_name': 'Linux %(branch)s',
@@ -264,7 +264,7 @@ PLATFORM_VARS = {
                 'reboot_command': ['scripts/external_tools/count_and_reboot.py',
                                    '-f', '../reboot_count.txt', '-n', '1', '-z'],
                 'script_timeout': 3 * 3600,
-                'script_maxtime': 5.5 * 3600,
+                'script_maxtime': int(5.5 * 3600),
             },
             # because non-unified platforms are defined at misc level,
             # we can not add a new platform in config.py for this but instead
@@ -363,7 +363,7 @@ PLATFORM_VARS = {
                 'reboot_command': ['scripts/external_tools/count_and_reboot.py',
                                    '-f', '../reboot_count.txt','-n', '1', '-z'],
                 'script_timeout': 3 * 3600,
-                'script_maxtime': 5.5 * 3600,
+                'script_maxtime': int(5.5 * 3600),
             },
 
             'product_name': 'firefox',
@@ -454,7 +454,7 @@ PLATFORM_VARS = {
                 'reboot_command': ['scripts/external_tools/count_and_reboot.py',
                                    '-f', '../reboot_count.txt','-n', '1', '-z'],
                 'script_timeout': 3 * 3600,
-                'script_maxtime': 5.5 * 3600,
+                'script_maxtime': int(5.5 * 3600),
             },
 
             'enable_nightly': True,
@@ -546,7 +546,7 @@ PLATFORM_VARS = {
                 'reboot_command': ['scripts/external_tools/count_and_reboot.py',
                                    '-f', '../reboot_count.txt','-n', '1', '-z'],
                 'script_timeout': 3 * 3600,
-                'script_maxtime': 5.5 * 3600,
+                'script_maxtime': int(5.5 * 3600(,
             },
 
             'enable_nightly': False,
@@ -671,7 +671,7 @@ PLATFORM_VARS = {
                 'reboot_command': ['scripts/external_tools/count_and_reboot.py',
                                    '-f', '../reboot_count.txt','-n', '1', '-z'],
                 'script_timeout': 3 * 3600,
-                'script_maxtime': 5.5 * 3600,
+                'script_maxtime': int(5.5 * 3600),
             },
 
             'consider_for_nightly': False,
@@ -744,7 +744,7 @@ PLATFORM_VARS = {
                 'reboot_command': ['scripts/external_tools/count_and_reboot.py',
                                    '-f', '../reboot_count.txt', '-n', '1', '-z'],
                 'script_timeout': 3 * 3600,
-                'script_maxtime': 5.5 * 3600,
+                'script_maxtime': int(5.5 * 3600),
             },
             # because non-unified platforms are defined at misc level,
             # we can not add a new platform in config.py for this but instead
@@ -876,7 +876,7 @@ PLATFORM_VARS = {
                     '-f', '../reboot_count.txt','-n', '1', '-z'
                 ],
                 'script_timeout': 3 * 3600,
-                'script_maxtime': 5.5 * 3600,
+                'script_maxtime': int(5.5 * 3600),
             },
             # because non-unified platforms are defined at misc level,
             # we can not add a new platform in config.py for this but instead
@@ -998,7 +998,7 @@ PLATFORM_VARS = {
                 'reboot_command': ['scripts/external_tools/count_and_reboot.py',
                                    '-f', '../reboot_count.txt','-n', '1', '-z'],
                 'script_timeout': 3 * 3600,
-                'script_maxtime': 5.5 * 3600,
+                'script_maxtime': int(5.5 * 3600),
             },
 
             'enable_nightly': False,
@@ -1096,7 +1096,7 @@ PLATFORM_VARS = {
                 'reboot_command': ['scripts/external_tools/count_and_reboot.py',
                                    '-f', '../reboot_count.txt','-n', '1', '-z'],
                 'script_timeout': 3 * 3600,
-                'script_maxtime': 5.5 * 3600,
+                'script_maxtime': int(5.5 * 3600),
             },
             # because non-unified platforms are defined at misc level,
             # we can not add a new platform in config.py for this but instead
@@ -1181,7 +1181,7 @@ PLATFORM_VARS = {
                 'reboot_command': ['scripts/external_tools/count_and_reboot.py',
                                    '-f', '../reboot_count.txt', '-n', '1', '-z'],
                 'script_timeout': 3 * 3600,
-                'script_maxtime': 5.5 * 3600,
+                'script_maxtime': int(5.5 * 3600),
             },
             # because non-unified platforms are defined at misc level,
             # we can not add a new platform in config.py for this but instead
@@ -1248,7 +1248,7 @@ PLATFORM_VARS = {
                     '-f', '../reboot_count.txt','-n', '1', '-z'
                 ],
                 'script_timeout': 3 * 3600,
-                'script_maxtime': 5.5 * 3600,
+                'script_maxtime': int(5.5 * 3600),
             },
             # because non-unified platforms are defined at misc level,
             # we can not add a new platform in config.py for this but instead
